+++ src/stores/trackInfo.ts
@@ -1,7 +1,7 @@
 import { writable } from 'svelte/store'
 import type { Writable } from 'svelte/store'
 import { methods } from '../stores/data'
-import type { Track, TrackID } from 'src/stores/libraryTypes'
+import type { Track, Image, TrackID } from 'src/stores/libraryTypes'
 
 export type TrackMD = {
   name: string
@@ -24,32 +24,32 @@ export type TrackMD = {
   comments: string
 }
 
-export const visible = writable(false)
 export const id: Writable<TrackID | null> = writable(null)
 export const track: Writable<Track | null> = writable(null)
-export function open(newId: TrackID) {
-  id.set(newId)
-  track.set(methods.getTrack(newId))
-  coverSrc.reset()
-  visible.set(true)
-  coverSrc.newFromTrackId(newId)
-}
+export const image = writable(null as null | Image)
 
-export const coverSrc = (() => {
-  const { set, subscribe }: Writable<string | null> = writable(null)
+export const visible = (() => {
+  const store = writable(false)
   return {
-    async newFromTrackId(id: TrackID) {
-      try {
-        let buf = await methods.readCoverAsync(id)
-        let url = URL.createObjectURL(new Blob([buf], {}))
-        set(url)
-      } catch (e) {
-        set(null)
+    subscribe: store.subscribe,
+    open: (newId: string) => {
+      id.set(newId)
+      track.set(methods.getTrack(newId))
+      store.set(true)
+      methods.loadTags(newId)
+      loadImage()
+    },
+    set: (value: boolean) => {
+      if (value === false) {
+        store.set(false)
+        id.set(null)
+        track.set(null)
+        image.set(null)
       }
     },
-    reset() {
-      set(null)
-    },
-    subscribe,
   }
 })()
+
+export function loadImage() {
+  image.set(methods.getImage(0))
+}