+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Next
 - Add ability to delete tracks from library
 - Add `Cmd+U` / `View > Toggle Queue` to show/hide the queue
+- Add support for adding and replacing track artworks
 - Add more items to the Song menu
 - Add `Escape` shortcut for deselecting tracks
 - Add support selecting tracks via `Cmd+A`, `Shift+Up/Down`, `Alt+Up/Down`, `Shift+Alt+Up/Down`